4to commit

+++ js/carrito.js
@@ -76,20 +76,20 @@ function eliminarDelCarrito(e) {
         text: "Producto eliminado",
         duration: 3000,
         close: true,
-        gravity: "top", // `top` or `bottom`
-        position: "right", // `left`, `center` or `right`
-        stopOnFocus: true, // Prevents dismissing of toast on hover
+        gravity: "top",
+        position: "center",
+        stopOnFocus: true,
         style: {
-            background: "linear-gradient(to right, #4b33a8, #785ce9)",
+            background: "red",
             borderRadius: "2rem",
             textTransform: "uppercase",
             fontSize: ".75rem"
         },
         offset: {
-            x: '1.5rem', // horizontal axis - can be a number or a string indicating unity. eg: '2em'
-            y: '1.5rem' // vertical axis - can be a number or a string indicating unity. eg: '2em'
+            x: '1.5rem',
+            y: '1.5rem'
             },
-        onClick: function(){} // Callback after click
+        onClick: function(){}
     }).showToast();
 
     const idBoton = e.currentTarget.id;
@@ -131,8 +131,19 @@ function actualizarTotal() {
 botonComprar.addEventListener("click", comprarCarrito);
 function comprarCarrito() {
 
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+    Swal.fire({
+        title: 'Gracias por el pedido!',
+        icon: 'success',
+        html: `Se le enviará un mail con los detalles`,
+        showCancelButton: false,
+        focusConfirm: false,
+        confirmButtonText: 'Ok',
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+        }
+    })
     
     contenedorCarritoVacio.classList.add("disabled");
     contenedorCarritoProductos.classList.add("disabled");
